_otter_env_core.sh: do not clobber existing PERL5LIB
Particularly important for when Perl dependencies of Otter are installed
by an unprivileged user, in which case they are likely to end up in a
directory pointed to by said environment variable rather than one
configured in a different fashion.

Warning: it is important that we *prepend* Otter directories to
PERL5LIB rather than append them because it turns out that
having existing PERL5LIB point to an Ensembl installation confuses
the Otter client.